Asyncio_7.8/requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0